Introdução ao jUnit5/demo/src/main/java/lista/lista01/Pessoa.java
package lista.lista01;

public class Pessoa {
    private double peso;
    private double altura;

    public Pessoa(double peso, double altura) {
        if (peso <= 0) {
            throw new IllegalArgumentException("Peso deve ser maior que zero.");
        }
        if (altura <= 0) {
            throw new IllegalArgumentException("Altura deve ser maior que zero.");
        }
        this.peso = peso;
        this.altura = altura;
    }

    public double calcularIMC() {
        return peso / (altura * altura);
    }
}
